Move rules for consumption metering points
