+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Consumption/Rules/InvalidSettlementMethodRuleError.cs
@@ -12,11 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Consumption.Rules
 {
-    public class ScheduledMeterReadingDateNotAllowedRuleError : ValidationError
+    public class InvalidSettlementMethodRuleError : ValidationError
     {
+        public InvalidSettlementMethodRuleError(SettlementMethod settlementMethod)
+        {
+            SettlementMethod = settlementMethod;
+        }
+
+        public SettlementMethod SettlementMethod { get; }
     }
 }
