+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/Consumption/ScheduledMeterReadingDate.cs
@@ -14,7 +14,7 @@
 
 using System;
 using System.Collections.ObjectModel;
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption.Rules;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Consumption.Rules;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
 namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption
