+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/Errors/Converters/Create/Consumption/ScheduledMeterReadingDateNotAllowedRuleErrorConverter.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption.Rules;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Consumption.Rules;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors.Converters.Create.Consumption
 {
