+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/Errors/Converters/MeterReadingOccurenceInvalidValueErrorConverter.cs
@@ -14,7 +14,6 @@
 
 using System;
 using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Consumption.Rules;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors.Converters
 {
